refactor(icon-service): fix typings
