+++ packages/vuetify/types/services/icons.d.ts
@@ -1,8 +1,6 @@
+// Types
 import { Icons } from '../../src/services/icons'
-
-export interface VuetifyIconService {
-  new (options?: VuetifyIconOptions): Icons
-}
+import { Component } from 'vue'
 
 export interface VuetifyIconOptions {
   /**
@@ -14,39 +12,49 @@ export interface VuetifyIconOptions {
    * fa4: <a href="">FontAwesome 4</a> TODO: link
    */
   iconfont: 'md' | 'mdi' | 'fa' | 'fa4' // TODO: camelCase
-  values?: VuetifyIcons
+  values?: {
+    [key: string]: VuetifyIcon
+  }
 }
 
 export interface VuetifyIconSets {
   [key: string]: VuetifyIcons
 }
 
+
+export type VuetifyIconComponent = {
+  component: Component | string
+  props?: object
+}
+export type VuetifyIcon = string | VuetifyIconComponent
+
 export interface VuetifyIcons {
-  cancel?: string
-  clear?: string
-  close?: string
-  complete?: string
-  delete?: string // delete (e.g. v-chip close)
-  success?: string
-  info?: string
-  warning?: string
-  error?: string
-  prev?: string
-  next?: string
-  checkboxOn?: string
-  checkboxOff?: string
-  checkboxIndeterminate?: string
-  delimiter?: string // for carousel
-  sort?: string
-  expand?: string
-  menu?: string
-  subgroup?: string
-  dropdown?: string
-  radioOn?: string
-  radioOff?: string
-  edit?: string
-  ratingEmpty?: string
-  ratingFull?: string
-  ratingHalf?: string
-  loading?: string
+  [name: string]: VuetifyIcon
+
+  complete: VuetifyIcon
+  cancel: VuetifyIcon
+  close: VuetifyIcon
+  delete: VuetifyIcon
+  clear: VuetifyIcon
+  success: VuetifyIcon
+  info: VuetifyIcon
+  warning: VuetifyIcon
+  error: VuetifyIcon
+  prev: VuetifyIcon
+  next: VuetifyIcon
+  checkboxOn: VuetifyIcon
+  checkboxOff: VuetifyIcon
+  checkboxIndeterminate: VuetifyIcon
+  delimiter: VuetifyIcon
+  sort: VuetifyIcon
+  expand: VuetifyIcon
+  menu: VuetifyIcon
+  subgroup: VuetifyIcon
+  dropdown: VuetifyIcon
+  radioOn: VuetifyIcon
+  radioOff: VuetifyIcon
+  edit: VuetifyIcon
+  ratingEmpty: VuetifyIcon
+  ratingFull: VuetifyIcon
+  ratingHalf: VuetifyIcon
 }
